+++ eclient/users.go
@@ -88,7 +88,7 @@ func (c *EcomClient) GetUsers(ctx context.Context) ([]*UserResponse, error) {
 
 	var userContainer UserContainer
 	if err := json.NewDecoder(res.Body).Decode(&userContainer); err != nil {
-		return nil, errors.Wrapf(err, "json decode url", uri)
+		return nil, errors.Wrapf(err, "json decode url=%q", uri)
 	}
 	return userContainer.Data, nil
 }